Improved headlessness
Headless mode now prints "OK" if the command succeeds,
and "FAIL" if an OleoJmp is thrown.

`headless_graphics()' is called early from main.cc which creates
a fallback position. Better graphics like drivers like ncurses
can override the functionality.

Headless mode also contains a default bell command. It just
prints "BELL". It also saves the rest of the program having
to figure out if the display is set.

If I get rid of the X11 display, then the whole logic of selecting
a display will be greatly simplified, as headless will be
available as the default, and curses can be selected if
headless is not force.