Page Up/Down + Home now work (issue #13)
New behaviour:
- Page Up executes scroll-up
- Page Down executes scroll-down
- Home executes beginning-of-row, consistent with Excel

This was achieved by adding the following lines to init.c:
+  "bind-key generic-ansi scroll-up 5",
+  "bind-key generic-ansi scroll-down 6",
+  "bind-key generic-ansi beginning-of-row H",

Note that neoleo never executes the function kepypad(..., TRUE)
except when dealing with menus. This means that it has to
interpret the ANSI escape sequences itself.

It is unclear how reliable this is in the general case, but it was
possibly done to maintain compatability with other interfaces.

What this means is that io-curses.cc:_io_read_kbd() may stuff
several chars into the return buffer, even if only one key like
Home is pressed. Conceptually, it would be easier if the keypad
was enabled and the proper ncurses mecahnism was used. This would
presumably make cmd.cc:read_get_chr() simpler, assuming you
did not introduce more bugs trying to unknot it.

In order to discover the bind-key values that I needed, I enabled
the debugging infor that I put in  io-curses.cc:_io_read_kbd().
The buffer generated by keypresses is as follows:
Page Up :  27 91 53 126
Page Down: 27 91 54 126
Insert:    27 91 50 126
Delete:    27 91 51 126
End:       27 91 70
Home:      27 91 72

Clearly, the 3rd char in the buffer determines what was set. A
summary of the ASCII chart is as follows:
dec char
50    2
51    3
52    4
53    5
54    6
70    F
71    G
72    H

That's how I knew, for example, that I should set the Home key to
bind-key ... H

I do not know how neoleo actually does this translation yet. It
is superfluous for the purposes of making it work.

Parenthetically, the ncurses functions putwin() and getwin() can
be used to save and restore the state of the window.

Close:
https://github.com/blippy/neoleo/issues/13