save-spreadsheet command frame behaviour improved
When user does
	M-x save-spreadsheet
the command frame bindings have been improved. Left arrow, right arrow,
forward delete, backspace, home and end keys work as expected. The
commands that implement them are in init.c:

+  "# generic frame bindings",
+  "create-keymap generic-frame universal",
+  "bind-set generic-frame self-insert-command 0-9",
+  "bind-set generic-frame self-insert-command a-z",
+  "bind-set generic-frame self-insert-command A-Z",
+  "bind-set generic-frame self-insert-command .",
+  "bind-set generic-frame self-insert-command -",
+  "bind-set generic-frame self-insert-command _",
+  "bind-key generic-frame backward-char D", // left arrow
+  "bind-key generic-frame forward-char C", // right arrow
+  "bind-key generic-frame backward-delete-char ^?", // Backspace key
+  "bind-key generic-frame delete-char 3", // Delete key
+  "bind-key generic-frame end-of-line F", // End key
+  "bind-key generic-frame beginning-of-line H", // Home key
+  "",

Key bindings to the read-filename keymap were too encompassing, so
it was changed as follows:
-  "create-keymap read-filename read-string",
+  "create-keymap read-filename generic-frame",

Keymap read-string is probably too generic in its current form
anyway; but that will be the subject of a future commit.

The function cmd.c::get_argument() is a good function in which to
turn on debugging in order to understand what key presses are
being generated.