Improvements in headless mode
Editing cells is now a lot easier with the `i' command.
Spreadsheet can be saved using `w' command.

This should open the door to a lot more testability of neoleo.
I want to do refactoring, but it's dangerous to do that without
tests. These improvements in headless mode make it a lot
easier to guard against regressions.