Turned some structs into classes
This is a bit awkward, because the cell constructor or
desctructor is never called. It appears that lists.cc is
responsible for creating cells. However, it just creates memory
of a certain size for a cell, rather than calling a class
constructor.

So doing things the right way is going to be a problem.