Fixed standards non-compliance in erase()
io-edit.cc:erase() used strcpy() to shorten a string. The
standards say that the src and dest may not overlap. But in
this case, they do; although the behaviour is correct.
The address sanitizer noticed the con-compliance, and threw
an error.

This commit rewrites the code so that that neoleo doesn't barf.