Config fixes
We now seem to be able to do a basic make dist && make distcheck

Added a python file to aur directory so that we can perform
automated builds better. Bash scripts are a bit too clumsy, and
need to be superceded.

This should pave the way for a more automated release process,
although I doubt it can be fully automated.