Added string function concata, removed concat
`concata' just concatenates two strings. I use it as a
"test function" for improving the encapsulation of functions.

concata is wrapped in a function wrapfunc(), which now acts as a
pinch point for that pesky obstack.

See stringo.cc for more details.

`concat' was removed because I never used it. It's a complex
function, so I just decided to eliminate it.