Experimenting with lexing
I'm wondering whether it's a good idea to using the existing lexer
to byte-encode the input. Generating the formula should be easier.
The only problem is evaluation, for which I could use a slight
adaption of parse.yy.  Instead of parsing raw text, it parses
the byte-stream. What's more, it can evaluate it straight away.

node.h also gets a small redefinition.