Working on alt cells
I am re-working the cells from lists.cc to alt-cells.cc. lists.cc
is such a nightmare to work with that it's easier just to rewrite it.
The code it 1/6 the size so far, although it's likely to double.

You can toggle between the legacy version and new version by tinkering
with
	lists.h:#define USE_ALT_CELLS
Leave it as defined to use the new version. Comment it out to use
the legacy version.

WARNING: It crashes as-is. The code is much simpler though!