Fixed C-l and C-r
Aligning cell left and right caused crashes. There was an
erroneous free(). I think this bug must have been there from
the beginning.

The problem is that sometimes the expanded_prompt is alloc'd,
and sometimes it uses a pointer from someone else. So you've
the problem that you don't know if the pointer really requires
free'ing. This is one of the many resource-sharing conundrums
in the code.

What's "special" about C-l and C-r is that they were defined
as macros, and they don't need any input from the user. This
combo may be causing things to go awry.

I have now eliminate them as macros, and declared two new
functions called
	set-cell-alignment-left
and
	set-cell-alignment-right
which don't take any arguments.

This wins that particular battle, but the war still goes on,
of course.

This fixes https://github.com/blippy/neoleo/issues/39