Fixed bug in row removal and insertion
When I went over to the fancy sheet version, there were some
problems. When you insert a row, you have to bump thw rows
in the right order, as you get over-writing.

Also, you need to account for ranges which haven't had their
references set properly.

One problem is that when cell parsing is called, it
hardcodes for cell position, even when the formula makes an
absolute reference. So its relative referencing can get
out-of-wack.

The "obvious" solution would be to not put the relative references
in as absolute references. Good idea, but this creates a problem
for eval.

Column insertion and removal has not yet been addressed, let
alone more complicated moving and copying.