Added prec_cells to cells class
Currently only works for ranges, but prec_cells contains all
the cells that a formula points to. This will eventually eliminate
range bugs on things like SUM, where a row is inserted.

A lot of stuff in ref.cc should be vastly simplified over time
as we make use of C++ features.